+++ test/validate_test.js
@@ -19,7 +19,7 @@ suite('api/validate', function() {
     method:   'post',
     route:    '/test-input',
     name:     'testInputValidate',
-    input:    'http://localhost:4321/test-schema.json',
+    input:    'test-schema.json',
     title:    'Test End-Point',
     description:  'Place we can call to test something',
   }, function(req, res) {
@@ -31,7 +31,7 @@ suite('api/validate', function() {
     method:   'get',
     route:    '/test-output',
     name:     'testInputValidOutputValidate',
-    output:   'http://localhost:4321/test-schema.json',
+    output:   'test-schema.json',
     title:    'Test End-Point',
     description:  'Place we can call to test something',
   }, function(req, res) {
@@ -43,7 +43,7 @@ suite('api/validate', function() {
     method:   'get',
     route:    '/test-invalid-output',
     name:     'testInputInvalidOutputValidate',
-    output:   'http://localhost:4321/test-schema.json',
+    output:   'test-schema.json',
     title:    'Test End-Point',
     description:  'Place we can call to test something',
   }, function(req, res) {
@@ -55,7 +55,7 @@ suite('api/validate', function() {
     method:   'post',
     route:    '/test-skip-input-validation',
     name:     'testInputSkipInputValidation',
-    input:    'http://localhost:4321/test-schema.json',
+    input:    'test-schema.json',
     skipInputValidation: true,
     title:    'Test End-Point',
     description:  'Place we can call to test something',
@@ -68,7 +68,7 @@ suite('api/validate', function() {
     method:   'get',
     route:    '/test-skip-output-validation',
     name:     'testOutputSkipOutputValidation',
-    output:    'http://localhost:4321/test-schema.json',
+    output:    'test-schema.json',
     skipOutputValidation: true,
     title:    'Test End-Point',
     description:  'Place we can call to test something',
